Back/app/TypeAppointment/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import *
from rest_framework import filters
from django_filters.rest_framework import DjangoFilterBackend
from .serializers import *
from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework_simplejwt.authentication import JWTAuthentication

# Create your views here.
class TypeAppointmentViewSet(viewsets.ModelViewSet):
    queryset = TypeAppointment.objects.all()
    serializer_class = TypeAppointmentSerializers
    permission_classes = [IsAuthenticated]
    authentication_classes = ([JWTAuthentication])
    
    filter_backends = [
        DjangoFilterBackend,
        filters.SearchFilter,  
        filters.OrderingFilter,
    ]
    
    filterset_fields = ('__all__')
    search_fields = ('id', 'nameAppointment', 'cost')
    ordering_fields = ('__all__')