Back/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-model-utils==4.5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.4
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0